refined comments from review part 1
